+++ files/default/tests/minitest/support/shibboleth-sp_iis_test.rb
@@ -1,7 +1,0 @@
-require "minitest/autorun"
-
-describe_recipe "shibboleth-sp::iis" do
-  include MiniTest::Chef::Assertions
-  include MiniTest::Chef::Context
-  include MiniTest::Chef::Resources
-end